--- a/cantools/database/can/database.py
+++ b/cantools/database/can/database.py
@@ -1,5 +1,6 @@
 import logging
 from typing import (
+    ByteString,
     Dict,
     List,
     Optional,
@@ -446,7 +447,7 @@ class Database:
 
     def decode_message(self,
                        frame_id_or_name: Union[int, str],
-                       data: bytes,
+                       data: ByteString,
                        decode_choices: bool = True,
                        scaling: bool = True,
                        decode_containers: bool = False,