… cantools/subparsers/__utils__.py
… cantools/typechecking.py
… tests/test_database.py