@zariiii9003 2023-05-11
> as an additional bonus, this PR also adds "inverse choices dicts", which should significantly increase encoding performance for messages which contain many signals that exhibit a lot of choices. @zariiii9003: would be great if you gave this a spin with your DBC file...


Sure. I'm not sure how many choices that message actually has, but decoding with `decode_choices=True` seems significantly faster. The other measurements are a little slower, probably because of additional function calls. I tested with python 3.11, the function call penalty might be higher with earlier python versions.
Overall, the performance looks fine to me.


### With `decode_choices`
Script:
```python3
unpacked = {signal.name: min(0x40, signal.maximum) for signal in message.signals}
packed = bytes([random.randint(0, 255) for _ in range(message.length)])

print(f"{len(message.signals)=}")
n = 10_000
dt = timeit.repeat('data = message.encode(unpacked, strict=False, scaling=True)', globals=globals(), repeat=5, number=n)
print(f"encode time: {min(dt)/n * 1e6} us")
dt = timeit.repeat('message.decode(packed, decode_choices=True, scaling=True)', globals=globals(), repeat=5, number=n)
print(f"decode time: {min(dt)/n * 1e6} us")
```

master:
```
len(message.signals)=39
encode time: 12.071810000270489 us
decode time: 18.893299999763258 us
```

field_to_signal:
```
len(message.signals)=39
encode time: 13.660729999901378 us
decode time: 7.343199999741046 us
```

### Without `decode_choices`
Script:
```python3
unpacked = {signal.name: min(0x40, signal.maximum) for signal in message.signals}
packed = bytes([random.randint(0, 255) for _ in range(message.length)])

print(f"{len(message.signals)=}")
n = 10_000
dt = timeit.repeat('data = message.encode(unpacked, strict=False, scaling=True)', globals=globals(), repeat=5, number=n)
print(f"encode time: {min(dt)/n * 1e6} us")
dt = timeit.repeat('message.decode(packed, decode_choices=False, scaling=True)', globals=globals(), repeat=5, number=n)
print(f"decode time: {min(dt)/n * 1e6} us")
```

master:
```
len(message.signals)=39
encode time: 12.39417000033427 us
decode time: 6.245800000033341 us
```

field_to_signal:
```
len(message.signals)=39
encode time: 14.142059999721823 us
decode time: 6.78241000023263 us
```

### Without `scaling`
Script:
```python3
unpacked = {signal.name: min(0x40, signal.maximum) for signal in message.signals}
packed = bytes([random.randint(0, 255) for _ in range(message.length)])

print(f"{len(message.signals)=}")
n = 10_000
dt = timeit.repeat('data = message.encode(unpacked, strict=False, scaling=False)', globals=globals(), repeat=5, number=n)
print(f"encode time: {min(dt)/n * 1e6} us")
dt = timeit.repeat('message.decode(packed, decode_choices=False, scaling=False)', globals=globals(), repeat=5, number=n)
print(f"decode time: {min(dt)/n * 1e6} us")
```

master:
```
len(message.signals)=39
encode time: 8.633569999801693 us
decode time: 4.039819999889005 us
```

field_to_signal:
```
len(message.signals)=39
encode time: 10.973010000088834 us
decode time: 4.345120000289171 us
```