@andlaus 2023-05-15
> Data/Signal is also used in utils.create_encode_decode_formats. You could adapt those names, too.

good point. I hope to have captured the non-diagnostic code paths [here](https://github.com/cantools/cantools/pull/561/commits/ef775516497bf43609b2ee971db62caa17af6492)